+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - When converting girder images locally, prefer mount paths ([886](../../pull/886))
 - Store the id of job results for easier post-job work ([887](../../pull/887))
 - Harden style compositing of partial tiles ([889](../../pull/889))
+- Cache read_tiff calls to speed up restyling ([891](../../pull/891))
 
 ### Changes
 - Be more consistent in source class name attribute assignment ([884](../../pull/884))